+++ src/kaplay.ts
@@ -25,6 +25,10 @@ import {
     drawTriangle,
     drawUnscaled,
     drawUVQuad,
+    beginPicture,
+    appendToPicture,
+    endPicture,
+    drawPicture,
     flush,
     formatText,
     FrameBuffer,
@@ -42,6 +46,8 @@ import {
     setBackground,
     updateViewport,
     width,
+    Picture,
+    drawCanvas,
 } from "./gfx";
 
 import {
@@ -164,6 +170,7 @@ import {
 
 import {
     BlendMode,
+    type Canvas,
     type Debug,
     type GameObj,
     type KAPLAYCtx,
@@ -502,7 +509,7 @@ const kaplay = <
         LCEvents.AfterUpdate,
     ]);
 
-    function makeCanvas(w: number, h: number) {
+    function makeCanvas(w: number, h: number): Canvas {
         const fb = new FrameBuffer(ggl, w, h);
 
         return {
@@ -1512,6 +1519,10 @@ const kaplay = <
         drawFormattedText,
         drawMasked,
         drawSubtracted,
+        beginPicture,
+        appendToPicture,
+        endPicture,
+        drawPicture,
         pushTransform,
         popTransform,
         pushTranslate: pushTranslateV,
@@ -1520,6 +1531,8 @@ const kaplay = <
         pushMatrix,
         usePostEffect,
         makeCanvas,
+        drawCanvas,
+        Picture,
         // debug
         debug,
         // scene