+++ CHANGELOG.md
@@ -37,6 +37,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Added LCEvents enum for identify different lifecycle events.
 - Moved camera to the shader - @mflerackers
 - Blend mode - @mflerackers
+- Picture API - @mflerackers
+- drawCanvas - @mflerackers
 - Added `video()` component - @mflerackers
 
 ### Changed